+++ app/src/main/res/drawable/ic_splash_screen.xml
@@ -1,0 +1,89 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="595.3dp"
+    android:height="841.9dp"
+    android:viewportWidth="595.3"
+    android:viewportHeight="841.9">
+  <path
+      android:pathData="M13.2,100.9h574.1v573.8h-574.1z"
+      android:fillColor="#FFFFFF"/>
+  <group>
+    <clip-path
+        android:pathData="M0,26h595.3v666h-595.3z"/>
+    <path
+        android:pathData="M390.9,350l-1.4,0.8c-3.3,-1.8 -6.8,-3.2 -10.4,-4.3l-0.6,-7.4C382,343.8 386.5,347.2 390.9,350z"
+        android:fillColor="#ED1B33"/>
+    <path
+        android:pathData="M423.1,368l-20.7,-24l-0.5,-9.1l-2.3,-2.1c-9.2,-8.3 -20.7,-12.1 -26.9,-13.7c-1.7,-3.5 -4.8,-8.4 -10.6,-13.3l-8.9,-7.7l-3.5,11.2c-0.3,1 -2.9,9.7 -1.1,19.2l0.1,0.7c-6.3,3.5 -17.6,8 -33.9,8c-1.1,0 -2.3,0 -3.5,-0.1l-19.8,-0.8l11.7,11.6l-7.8,-1.8c-28.7,-23.3 -76.3,-46 -154.9,-54.9l-19.4,-2.2l12.7,14.8c1.2,1.4 26.5,30.5 61.1,51.7c-15.4,0.6 -31.8,2.8 -48.7,6.5l-8.9,2l3.5,8.4c0.6,1.5 14.4,33.4 58.7,40.6c-10.7,5.4 -20.7,12.5 -29.8,21.3l-7,6.7l8.1,5.2c0.8,0.5 19.5,12.4 43.3,12.4c5.1,0 10,-0.5 14.8,-1.6c-0.8,1.6 -3.7,27.4 -3.7,27.4c7.6,-36.4 35.7,-52.5 58.3,-57.8c-32.9,1.7 -42.4,17.8 -42.4,17.8c-32.5,17.1 -66.1,-4.7 -66.1,-4.7c53.6,-51.5 120,-25.5 120,-25.5c-39,-18 -75.8,-7.6 -75.8,-7.6c-55.6,0.6 -71.3,-37.3 -71.3,-37.3c98.6,-21.6 144.9,16.2 159.2,31.7c-39.9,-46.5 -92.1,-43.4 -92.1,-43.4c-42.1,-20.3 -75.4,-59 -75.4,-59C317.9,319 331.1,411.7 331.1,411.7c-1,-27.1 -29,-56.1 -29,-56.1c10.5,2.5 21.6,-2.9 21.6,-2.9c-7.2,-2.3 -12.7,-7.8 -12.7,-7.8c32.3,1.3 48.9,-13.9 48.9,-13.9c-0.8,-2 -3.8,-3.7 -3.8,-3.7c-1.5,-7.9 0.9,-15.5 0.9,-15.5c8.7,7.5 10,14.2 10,14.2s15.9,2.3 27.4,12.7l0.5,8.6l25.1,29l4.4,-0.5l-3.3,13.6l-17.7,1.9c0,0 -4,-18 -31.1,-14.8c0,0 -14.3,-3.4 -11.3,-19.2c0,0 -20.5,49.5 16.1,84.8l20.3,14.7c0,0 -40.6,-36.7 -34,-76c2.4,1.6 4.9,2.6 7.2,3.2l1.3,0.3l1.3,-0.2c1.8,-0.2 3.6,-0.3 5.3,-0.3c15.2,0 17.4,9.1 17.4,9.2l1.5,6.7l29.9,-3.3l7.1,-29.4L423.1,368z"
+        android:fillColor="#ED1B33"/>
+  </group>
+  <group>
+    <clip-path
+        android:pathData="M0,26h595.3v666h-595.3z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M273.2,448.5c0,0.5 -0.3,0.8 -0.8,0.8h-10v15.4c0,0.5 -0.3,0.8 -0.8,0.8s-0.8,-0.3 -0.8,-0.8v-15.4h-10c-0.5,0 -0.8,-0.3 -0.8,-0.8s0.3,-0.8 0.8,-0.8h21.6C272.9,447.7 273.2,448 273.2,448.5z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M275.7,447.9c0.5,0 0.8,0.3 0.8,0.8v16.2c0,0.5 -0.3,0.7 -0.8,0.7s-0.7,-0.2 -0.7,-0.7v-16.2C275,448.1 275.2,447.9 275.7,447.9z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M300.6,452.4c-0.1,0.1 -0.3,0.2 -0.5,0.2s-0.4,-0.1 -0.5,-0.2c-1,-1.1 -2.3,-2 -3.8,-2.6c-1.6,-0.6 -3.2,-1 -5,-1c-2.9,0 -5.4,0.8 -7.4,2.3c-2.1,1.5 -3.1,3.4 -3.1,5.5s1,4 3.1,5.5s4.5,2.3 7.4,2.3c1.8,0 3.5,-0.3 5,-1c1.6,-0.6 2.8,-1.5 3.8,-2.6c0.1,-0.2 0.3,-0.2 0.5,-0.2s0.4,0.1 0.5,0.2c0.4,0.3 0.4,0.7 0.1,1.1c-1.1,1.3 -2.5,2.3 -4.3,3s-3.6,1.1 -5.6,1.1c-3.3,0 -6.1,-0.9 -8.4,-2.6c-2.4,-1.8 -3.6,-4 -3.6,-6.6s1.2,-4.8 3.6,-6.6c2.3,-1.8 5.1,-2.6 8.4,-2.6c2,0 3.9,0.4 5.6,1.1s3.2,1.7 4.3,3C301,451.8 300.9,452.1 300.6,452.4z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M326.2,464.8c0,0.5 -0.3,0.8 -0.8,0.8c-0.2,0 -0.4,-0.1 -0.5,-0.2L314.5,455l-9.9,5v4.9c0,0.5 -0.3,0.7 -0.8,0.7s-0.8,-0.2 -0.8,-0.7v-16.2c0,-0.5 0.3,-0.8 0.8,-0.8s0.8,0.3 0.8,0.8v9.6l20.5,-10.3c0.3,-0.1 0.5,-0.1 0.7,0s0.4,0.4 0.4,0.6c0,0.3 -0.1,0.5 -0.4,0.6l-9.9,4.9l10.1,10.1C326.1,464.4 326.2,464.6 326.2,464.8z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M329.5,457.3v6.6h20.8c0.5,0 0.8,0.3 0.8,0.8s-0.3,0.8 -0.8,0.8H328v-17.7h22.4c0.5,0 0.8,0.3 0.8,0.8s-0.3,0.8 -0.8,0.8h-20.8v6.6h20.8c0.5,0 0.8,0.3 0.8,0.8s-0.3,0.8 -0.8,0.8h-20.9V457.3z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M375.8,448.5c0,0.5 -0.3,0.8 -0.8,0.8h-10v15.4c0,0.5 -0.3,0.8 -0.8,0.8s-0.8,-0.3 -0.8,-0.8v-15.4h-10c-0.5,0 -0.8,-0.3 -0.8,-0.8s0.3,-0.8 0.8,-0.8H375C375.6,447.7 375.8,448 375.8,448.5z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M315.6,478.6v18.7c0,0.5 -0.2,0.7 -0.7,0.7s-0.8,-0.2 -0.8,-0.7v-13.7l-10,15.1L294,483.5v13.7c0,0.5 -0.2,0.7 -0.7,0.7s-0.8,-0.2 -0.8,-0.7v-18.7l11.6,17.3L315.6,478.6z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M318.8,489.7v6.6h20.8c0.5,0 0.8,0.3 0.8,0.8s-0.3,0.8 -0.8,0.8h-22.4v-17.7h22.4c0.5,0 0.8,0.3 0.8,0.8s-0.3,0.8 -0.8,0.8h-20.8v6.6h20.8c0.5,0 0.8,0.3 0.8,0.8s-0.3,0.8 -0.8,0.8h-20.8V489.7z"/>
+  </group>
+  <group>
+    <clip-path
+        android:pathData="M0,0h595.3v666h-595.3z"/>
+    <path
+        android:pathData="M-64,20a157,93 0,1 0,314 0a157,93 0,1 0,-314 0z"
+        android:fillColor="#ED1D26"/>
+  </group>
+  <group>
+    <clip-path
+        android:pathData="M0,0h595.3v666h-595.3z"/>
+    <path
+        android:pathData="M226,-26a104.2,113 0,1 0,208.4 0a104.2,113 0,1 0,-208.4 0z"
+        android:fillColor="#ED1D26"/>
+  </group>
+  <group>
+    <clip-path
+        android:pathData="M0,0h595.3v666h-595.3z"/>
+    <path
+        android:pathData="M320,39a199.5,74 0,1 0,399 0a199.5,74 0,1 0,-399 0z"
+        android:fillColor="#ED1D26"/>
+  </group>
+  <group>
+    <clip-path
+        android:pathData="M0,176h595.3v666h-595.3z"/>
+    <path
+        android:pathData="M-104,896.5a121.5,112.5 0,1 0,243 0a121.5,112.5 0,1 0,-243 0z"
+        android:fillColor="#ED1D26"/>
+  </group>
+  <group>
+    <clip-path
+        android:pathData="M0,176h595.3v666h-595.3z"/>
+    <path
+        android:pathData="M309,861.5a176,117.5 0,1 0,352 0a176,117.5 0,1 0,-352 0z"
+        android:fillColor="#ED1D26"/>
+  </group>
+  <group>
+    <clip-path
+        android:pathData="M0,176h595.3v666h-595.3z"/>
+    <path
+        android:pathData="M59,857.5a176,84.5 0,1 0,352 0a176,84.5 0,1 0,-352 0z"
+        android:fillColor="#ED1D26"/>
+  </group>
+</vector>
